CI: ci.py refactoring
