+++ tests/ci/ci.py
@@ -3,23 +3,25 @@ import concurrent.futures
 import json
 import logging
 import os
-import random
 import re
 import subprocess
 import sys
-import time
-from copy import deepcopy
-from dataclasses import asdict, dataclass
-from enum import Enum
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Sequence, Set, Tuple, Union
+from typing import Any, Dict, List, Optional
 
 import docker_images_helper
 import upload_result_helper
 from build_check import get_release_or_pr
-from ci_config import CI_CONFIG, Build, CILabels, CIStages, JobNames, StatusNames
+from ci_config import (
+    CI_CONFIG,
+    Build,
+    CILabels,
+    CIStages,
+    JobNames,
+    StatusNames,
+)
 from ci_metadata import CiMetadata
-from ci_utils import GHActions, is_hex, normalize_string
+from ci_utils import GHActions, normalize_string
 from clickhouse_helper import (
     CiLogsCredentials,
     ClickHouseHelper,
@@ -36,18 +38,14 @@ from commit_status_helper import (
     post_commit_status,
     set_status_comment,
 )
-from digest_helper import DockerDigester, JobDigester
+from digest_helper import DockerDigester
 from env_helper import (
     CI,
     GITHUB_JOB_API_URL,
     GITHUB_REPOSITORY,
     GITHUB_RUN_ID,
-    GITHUB_RUN_URL,
     REPO_COPY,
-    REPORT_PATH,
-    S3_BUILDS_BUCKET,
     TEMP_PATH,
-    CI_CONFIG_PATH,
 )
 from get_robot_token import get_best_robot_token
 from git_helper import GIT_PREFIX, Git
@@ -58,957 +56,17 @@ from report import ERROR, FAILURE, PENDING, SUCCESS, BuildResult, JobReport, Tes
 from s3_helper import S3Helper
 from stopwatch import Stopwatch
 from tee_popen import TeePopen
+from ci_cache import CiCache
+from ci_settings import CiSettings
 from version_helper import get_version_from_repo
 
 # pylint: disable=too-many-lines
-
-
-@dataclass
-class PendingState:
-    updated_at: float
-    run_url: str
-
-
-class CiCache:
-    """
-    CI cache is a bunch of records. Record is a file stored under special location on s3.
-    The file name has a format:
-
-        <RECORD_TYPE>_[<ATTRIBUTES>]--<JOB_NAME>_<JOB_DIGEST>_<BATCH>_<NUM_BATCHES>.ci
-
-    RECORD_TYPE:
-        SUCCESSFUL - for successful jobs
-        PENDING - for pending jobs
-
-    ATTRIBUTES:
-        release - for jobs being executed on the release branch including master branch (not a PR branch)
-    """
-
-    _S3_CACHE_PREFIX = "CI_cache_v1"
-    _CACHE_BUILD_REPORT_PREFIX = "build_report"
-    _RECORD_FILE_EXTENSION = ".ci"
-    _LOCAL_CACHE_PATH = Path(TEMP_PATH) / "ci_cache"
-    _ATTRIBUTE_RELEASE = "release"
-    # divider symbol 1
-    _DIV1 = "--"
-    # divider symbol 2
-    _DIV2 = "_"
-    assert _DIV1 != _DIV2
-
-    class RecordType(Enum):
-        SUCCESSFUL = "successful"
-        PENDING = "pending"
-        FAILED = "failed"
-
-    @dataclass
-    class Record:
-        record_type: "CiCache.RecordType"
-        job_name: str
-        job_digest: str
-        batch: int
-        num_batches: int
-        release_branch: bool
-        file: str = ""
-
-        def to_str_key(self):
-            """other fields must not be included in the hash str"""
-            return "_".join(
-                [self.job_name, self.job_digest, str(self.batch), str(self.num_batches)]
-            )
-
-    class JobType(Enum):
-        DOCS = "DOCS"
-        SRCS = "SRCS"
-
-        @classmethod
-        def is_docs_job(cls, job_name: str) -> bool:
-            return job_name == JobNames.DOCS_CHECK
-
-        @classmethod
-        def is_srcs_job(cls, job_name: str) -> bool:
-            return not cls.is_docs_job(job_name)
-
-        @classmethod
-        def get_type_by_name(cls, job_name: str) -> "CiCache.JobType":
-            res = cls.SRCS
-            if cls.is_docs_job(job_name):
-                res = cls.DOCS
-            elif cls.is_srcs_job(job_name):
-                res = cls.SRCS
-            else:
-                assert False
-            return res
-
-    def __init__(
-        self,
-        s3: S3Helper,
-        job_digests: Dict[str, str],
-    ):
-        self.s3 = s3
-        self.job_digests = job_digests
-        self.cache_s3_paths = {
-            job_type: f"{self._S3_CACHE_PREFIX}/{job_type.value}-{self._get_digest_for_job_type(self.job_digests, job_type)}/"
-            for job_type in self.JobType
-        }
-        self.s3_record_prefixes = {
-            record_type: record_type.value for record_type in self.RecordType
-        }
-        self.records: Dict["CiCache.RecordType", Dict[str, "CiCache.Record"]] = {
-            record_type: {} for record_type in self.RecordType
-        }
-
-        self.cache_updated = False
-        self.cache_data_fetched = True
-        if not self._LOCAL_CACHE_PATH.exists():
-            self._LOCAL_CACHE_PATH.mkdir(parents=True, exist_ok=True)
-
-    def _get_digest_for_job_type(
-        self, job_digests: Dict[str, str], job_type: JobType
-    ) -> str:
-        if job_type == self.JobType.DOCS:
-            res = job_digests[JobNames.DOCS_CHECK]
-        elif job_type == self.JobType.SRCS:
-            # any build type job has the same digest - pick up Build.PACKAGE_RELEASE or Build.PACKAGE_ASAN as a failover
-            # Build.PACKAGE_RELEASE may not exist in the list if we have reduced CI pipeline
-            if Build.PACKAGE_RELEASE in job_digests:
-                res = job_digests[Build.PACKAGE_RELEASE]
-            elif Build.PACKAGE_ASAN in job_digests:
-                # failover, if failover does not work - fix it!
-                res = job_digests[Build.PACKAGE_ASAN]
-            else:
-                assert False, "BUG, no build job in digest' list"
-        else:
-            assert False, "BUG, New JobType? - please update func"
-        return res
-
-    def _get_record_file_name(
-        self,
-        record_type: RecordType,
-        job_name: str,
-        batch: int,
-        num_batches: int,
-        release_branch: bool,
-    ) -> str:
-        prefix = self.s3_record_prefixes[record_type]
-        prefix_extended = (
-            self._DIV2.join([prefix, self._ATTRIBUTE_RELEASE])
-            if release_branch
-            else prefix
-        )
-        assert self._DIV1 not in job_name, f"Invalid job name {job_name}"
-        job_name = self._DIV2.join(
-            [job_name, self.job_digests[job_name], str(batch), str(num_batches)]
-        )
-        file_name = self._DIV1.join([prefix_extended, job_name])
-        file_name += self._RECORD_FILE_EXTENSION
-        return file_name
-
-    def _get_record_s3_path(self, job_name: str) -> str:
-        return self.cache_s3_paths[self.JobType.get_type_by_name(job_name)]
-
-    def _parse_record_file_name(
-        self, record_type: RecordType, file_name: str
-    ) -> Optional["CiCache.Record"]:
-        # validate filename
-        if (
-            not file_name.endswith(self._RECORD_FILE_EXTENSION)
-            or not len(file_name.split(self._DIV1)) == 2
-        ):
-            print("ERROR: wrong file name format")
-            return None
-
-        file_name = file_name.removesuffix(self._RECORD_FILE_EXTENSION)
-        release_branch = False
-
-        prefix_extended, job_suffix = file_name.split(self._DIV1)
-        record_type_and_attribute = prefix_extended.split(self._DIV2)
-
-        # validate filename prefix
-        failure = False
-        if not 0 < len(record_type_and_attribute) <= 2:
-            print("ERROR: wrong file name prefix")
-            failure = True
-        if (
-            len(record_type_and_attribute) > 1
-            and record_type_and_attribute[1] != self._ATTRIBUTE_RELEASE
-        ):
-            print("ERROR: wrong record attribute")
-            failure = True
-        if record_type_and_attribute[0] != self.s3_record_prefixes[record_type]:
-            print("ERROR: wrong record type")
-            failure = True
-        if failure:
-            return None
-
-        if (
-            len(record_type_and_attribute) > 1
-            and record_type_and_attribute[1] == self._ATTRIBUTE_RELEASE
-        ):
-            release_branch = True
-
-        job_properties = job_suffix.split(self._DIV2)
-        job_name, job_digest, batch, num_batches = (
-            self._DIV2.join(job_properties[:-3]),
-            job_properties[-3],
-            int(job_properties[-2]),
-            int(job_properties[-1]),
-        )
-
-        if not is_hex(job_digest):
-            print("ERROR: wrong record job digest")
-            return None
-
-        record = self.Record(
-            record_type,
-            job_name,
-            job_digest,
-            batch,
-            num_batches,
-            release_branch,
-            file="",
-        )
-        return record
-
-    def print_status(self):
-        for record_type in self.RecordType:
-            GHActions.print_in_group(
-                f"Cache records: [{record_type}]", list(self.records[record_type])
-            )
-        return self
-
-    @staticmethod
-    def dump_run_config(indata: Dict[str, Any]) -> None:
-        assert indata
-        assert CI_CONFIG_PATH
-        with open(CI_CONFIG_PATH, "w", encoding="utf-8") as json_file:
-            json.dump(indata, json_file, indent=2)
-
-    def update(self):
-        """
-        Pulls cache records from s3. Only records name w/o content.
-        """
-        for record_type in self.RecordType:
-            prefix = self.s3_record_prefixes[record_type]
-            cache_list = self.records[record_type]
-            for job_type in self.JobType:
-                path = self.cache_s3_paths[job_type]
-                records = self.s3.list_prefix(f"{path}{prefix}", S3_BUILDS_BUCKET)
-                records = [record.split("/")[-1] for record in records]
-                for file in records:
-                    record = self._parse_record_file_name(
-                        record_type=record_type, file_name=file
-                    )
-                    if not record:
-                        print(f"ERROR: failed to parse cache record [{file}]")
-                        continue
-                    if (
-                        record.job_name not in self.job_digests
-                        or self.job_digests[record.job_name] != record.job_digest
-                    ):
-                        # skip records we are not interested in
-                        continue
-
-                    if record.to_str_key() not in cache_list:
-                        cache_list[record.to_str_key()] = record
-                        self.cache_data_fetched = False
-                    elif (
-                        not cache_list[record.to_str_key()].release_branch
-                        and record.release_branch
-                    ):
-                        # replace a non-release record with a release one
-                        cache_list[record.to_str_key()] = record
-                        self.cache_data_fetched = False
-
-        self.cache_updated = True
-        return self
-
-    def fetch_records_data(self):
-        """
-        Pulls CommitStatusData for all cached jobs from s3
-        """
-        if not self.cache_updated:
-            self.update()
-
-        if self.cache_data_fetched:
-            # there are no records without fetched data - no need to fetch
-            return self
-
-        # clean up
-        for file in self._LOCAL_CACHE_PATH.glob("*.ci"):
-            file.unlink()
-
-        # download all record files
-        for job_type in self.JobType:
-            path = self.cache_s3_paths[job_type]
-            for record_type in self.RecordType:
-                prefix = self.s3_record_prefixes[record_type]
-                _ = self.s3.download_files(
-                    bucket=S3_BUILDS_BUCKET,
-                    s3_path=f"{path}{prefix}",
-                    file_suffix=self._RECORD_FILE_EXTENSION,
-                    local_directory=self._LOCAL_CACHE_PATH,
-                )
-
-        # validate we have files for all records and save file names meanwhile
-        for record_type in self.RecordType:
-            record_list = self.records[record_type]
-            for _, record in record_list.items():
-                record_file_name = self._get_record_file_name(
-                    record_type,
-                    record.job_name,
-                    record.batch,
-                    record.num_batches,
-                    record.release_branch,
-                )
-                assert (
-                    self._LOCAL_CACHE_PATH / record_file_name
-                ).is_file(), f"BUG. Record file must be present: {self._LOCAL_CACHE_PATH / record_file_name}"
-                record.file = record_file_name
-
-        self.cache_data_fetched = True
-        return self
-
-    def exist(
-        self,
-        record_type: "CiCache.RecordType",
-        job: str,
-        batch: int,
-        num_batches: int,
-        release_branch: bool,
-    ) -> bool:
-        if not self.cache_updated:
-            self.update()
-        record_key = self.Record(
-            record_type,
-            job,
-            self.job_digests[job],
-            batch,
-            num_batches,
-            release_branch,
-        ).to_str_key()
-        res = record_key in self.records[record_type]
-        if release_branch:
-            return res and self.records[record_type][record_key].release_branch
-        else:
-            return res
-
-    def push(
-        self,
-        record_type: "CiCache.RecordType",
-        job: str,
-        batches: Union[int, Sequence[int]],
-        num_batches: int,
-        status: Union[CommitStatusData, PendingState],
-        release_branch: bool = False,
-    ) -> None:
-        """
-        Pushes a cache record (CommitStatusData)
-        @release_branch adds "release" attribute to a record
-        """
-        if isinstance(batches, int):
-            batches = [batches]
-        for batch in batches:
-            record_file = self._LOCAL_CACHE_PATH / self._get_record_file_name(
-                record_type, job, batch, num_batches, release_branch
-            )
-            record_s3_path = self._get_record_s3_path(job)
-            if record_type == self.RecordType.SUCCESSFUL:
-                assert isinstance(status, CommitStatusData)
-                status.dump_to_file(record_file)
-            elif record_type == self.RecordType.FAILED:
-                assert isinstance(status, CommitStatusData)
-                status.dump_to_file(record_file)
-            elif record_type == self.RecordType.PENDING:
-                assert isinstance(status, PendingState)
-                with open(record_file, "w", encoding="utf-8") as json_file:
-                    json.dump(asdict(status), json_file)
-            else:
-                assert False
-
-            _ = self.s3.upload_file(
-                bucket=S3_BUILDS_BUCKET,
-                file_path=record_file,
-                s3_path=record_s3_path + record_file.name,
-            )
-            record = self.Record(
-                record_type,
-                job,
-                self.job_digests[job],
-                batch,
-                num_batches,
-                release_branch,
-                file=record_file.name,
-            )
-            if (
-                record.release_branch
-                or record.to_str_key() not in self.records[record_type]
-            ):
-                self.records[record_type][record.to_str_key()] = record
-
-    def get(
-        self, record_type: "CiCache.RecordType", job: str, batch: int, num_batches: int
-    ) -> Optional[Union[CommitStatusData, PendingState]]:
-        """
-        Gets a cache record data for a job, or None if a cache miss
-        """
-
-        if not self.cache_data_fetched:
-            self.fetch_records_data()
-
-        record_key = self.Record(
-            record_type,
-            job,
-            self.job_digests[job],
-            batch,
-            num_batches,
-            release_branch=False,
-        ).to_str_key()
-
-        if record_key not in self.records[record_type]:
-            return None
-
-        record_file_name = self.records[record_type][record_key].file
-
-        res = CommitStatusData.load_from_file(
-            self._LOCAL_CACHE_PATH / record_file_name
-        )  # type: CommitStatusData
-
-        return res
-
-    def delete(
-        self,
-        record_type: "CiCache.RecordType",
-        job: str,
-        batch: int,
-        num_batches: int,
-        release_branch: bool,
-    ) -> None:
-        """
-        deletes record from the cache
-        """
-        raise NotImplementedError("Let's try make cache push-and-read-only")
-        # assert (
-        #     record_type == self.RecordType.PENDING
-        # ), "FIXME: delete is supported for pending records only"
-        # record_file_name = self._get_record_file_name(
-        #     self.RecordType.PENDING,
-        #     job,
-        #     batch,
-        #     num_batches,
-        #     release_branch=release_branch,
-        # )
-        # record_s3_path = self._get_record_s3_path(job)
-        # self.s3.delete_file_from_s3(S3_BUILDS_BUCKET, record_s3_path + record_file_name)
-
-        # record_key = self.Record(
-        #     record_type,
-        #     job,
-        #     self.job_digests[job],
-        #     batch,
-        #     num_batches,
-        #     release_branch=False,
-        # ).to_str_key()
-
-        # if record_key in self.records[record_type]:
-        #     del self.records[record_type][record_key]
-
-    def is_successful(
-        self, job: str, batch: int, num_batches: int, release_branch: bool
-    ) -> bool:
-        """
-        checks if a given job have already been done successfully
-        """
-        return self.exist(
-            self.RecordType.SUCCESSFUL, job, batch, num_batches, release_branch
-        )
-
-    def is_failed(
-        self, job: str, batch: int, num_batches: int, release_branch: bool
-    ) -> bool:
-        """
-        checks if a given job have already been done with failure
-        """
-        return self.exist(
-            self.RecordType.FAILED, job, batch, num_batches, release_branch
-        )
-
-    def is_pending(
-        self, job: str, batch: int, num_batches: int, release_branch: bool
-    ) -> bool:
-        """
-        check pending record in the cache for a given job
-        @release_branch - checks that "release" attribute is set for a record
-        """
-        if self.is_successful(
-            job, batch, num_batches, release_branch
-        ) or self.is_failed(job, batch, num_batches, release_branch):
-            return False
-
-        return self.exist(
-            self.RecordType.PENDING, job, batch, num_batches, release_branch
-        )
-
-    def push_successful(
-        self,
-        job: str,
-        batch: int,
-        num_batches: int,
-        job_status: CommitStatusData,
-        release_branch: bool = False,
-    ) -> None:
-        """
-        Pushes a cache record (CommitStatusData)
-        @release_branch adds "release" attribute to a record
-        """
-        self.push(
-            self.RecordType.SUCCESSFUL,
-            job,
-            [batch],
-            num_batches,
-            job_status,
-            release_branch,
-        )
-
-    def push_failed(
-        self,
-        job: str,
-        batch: int,
-        num_batches: int,
-        job_status: CommitStatusData,
-        release_branch: bool = False,
-    ) -> None:
-        """
-        Pushes a cache record of type Failed (CommitStatusData)
-        @release_branch adds "release" attribute to a record
-        """
-        self.push(
-            self.RecordType.FAILED,
-            job,
-            [batch],
-            num_batches,
-            job_status,
-            release_branch,
-        )
-
-    def push_pending(
-        self, job: str, batches: List[int], num_batches: int, release_branch: bool
-    ) -> None:
-        """
-        pushes pending record for a job to the cache
-        """
-        pending_state = PendingState(time.time(), run_url=GITHUB_RUN_URL)
-        self.push(
-            self.RecordType.PENDING,
-            job,
-            batches,
-            num_batches,
-            pending_state,
-            release_branch,
-        )
-
-    def get_successful(
-        self, job: str, batch: int, num_batches: int
-    ) -> Optional[CommitStatusData]:
-        """
-        Gets a cache record (CommitStatusData) for a job, or None if a cache miss
-        """
-        res = self.get(self.RecordType.SUCCESSFUL, job, batch, num_batches)
-        assert res is None or isinstance(res, CommitStatusData)
-        return res
-
-    def delete_pending(
-        self, job: str, batch: int, num_batches: int, release_branch: bool
-    ) -> None:
-        """
-        deletes pending record from the cache
-        """
-        self.delete(self.RecordType.PENDING, job, batch, num_batches, release_branch)
-
-    def download_build_reports(self, file_prefix: str = "") -> List[str]:
-        """
-        not an ideal class for this method,
-        but let it be as we store build reports in CI cache directory on s3
-        and CiCache knows where exactly
-
-        @file_prefix allows filtering out reports by git head_ref
-        """
-        report_path = Path(REPORT_PATH)
-        report_path.mkdir(exist_ok=True, parents=True)
-        path = (
-            self._get_record_s3_path(Build.PACKAGE_RELEASE)
-            + self._CACHE_BUILD_REPORT_PREFIX
-        )
-        if file_prefix:
-            path += "_" + file_prefix
-        reports_files = self.s3.download_files(
-            bucket=S3_BUILDS_BUCKET,
-            s3_path=path,
-            file_suffix=".json",
-            local_directory=report_path,
-        )
-        return reports_files
-
-    def upload_build_report(self, build_result: BuildResult) -> str:
-        result_json_path = build_result.write_json(Path(TEMP_PATH))
-        s3_path = (
-            self._get_record_s3_path(Build.PACKAGE_RELEASE) + result_json_path.name
-        )
-        return self.s3.upload_file(
-            bucket=S3_BUILDS_BUCKET, file_path=result_json_path, s3_path=s3_path
-        )
-
-    def await_jobs(
-        self, jobs_with_params: Dict[str, Dict[str, Any]], is_release_branch: bool
-    ) -> Dict[str, List[int]]:
-        """
-        await pending jobs to be finished
-        @jobs_with_params - jobs to await. {JOB_NAME: {"batches": [BATCHES...], "num_batches": NUM_BATCHES}}
-        returns successfully finished jobs: {JOB_NAME: [BATCHES...]}
-        """
-        if not jobs_with_params:
-            return {}
-        poll_interval_sec = 300
-        # TIMEOUT * MAX_ROUNDS_TO_WAIT must be less than 6h (GH job timeout) with a room for rest RunConfig work
-        TIMEOUT = 3000  # 50 min
-        MAX_ROUNDS_TO_WAIT = 6
-        MAX_JOB_NUM_TO_WAIT = 3
-        await_finished: Dict[str, List[int]] = {}
-        round_cnt = 0
-        while (
-            len(jobs_with_params) > MAX_JOB_NUM_TO_WAIT
-            and round_cnt < MAX_ROUNDS_TO_WAIT
-        ):
-            round_cnt += 1
-            GHActions.print_in_group(
-                f"Wait pending jobs, round [{round_cnt}/{MAX_ROUNDS_TO_WAIT}]:",
-                list(jobs_with_params),
-            )
-            # this is initial approach to wait pending jobs:
-            # start waiting for the next TIMEOUT seconds if there are more than X(=4) jobs to wait
-            # wait TIMEOUT seconds in rounds. Y(=5) is the max number of rounds
-            expired_sec = 0
-            start_at = int(time.time())
-            while expired_sec < TIMEOUT and jobs_with_params:
-                time.sleep(poll_interval_sec)
-                self.update()
-                jobs_with_params_copy = deepcopy(jobs_with_params)
-                for job_name in jobs_with_params:
-                    num_batches = jobs_with_params[job_name]["num_batches"]
-                    job_config = CI_CONFIG.get_job_config(job_name)
-                    for batch in jobs_with_params[job_name]["batches"]:
-                        if self.is_pending(
-                            job_name,
-                            batch,
-                            num_batches,
-                            release_branch=is_release_branch
-                            and job_config.required_on_release_branch,
-                        ):
-                            continue
-                        print(
-                            f"Job [{job_name}_[{batch}/{num_batches}]] is not pending anymore"
-                        )
-
-                        # some_job_ready = True
-                        jobs_with_params_copy[job_name]["batches"].remove(batch)
-                        if not jobs_with_params_copy[job_name]["batches"]:
-                            del jobs_with_params_copy[job_name]
-
-                        if not self.is_successful(
-                            job_name,
-                            batch,
-                            num_batches,
-                            release_branch=is_release_branch
-                            and job_config.required_on_release_branch,
-                        ):
-                            print(
-                                f"NOTE: Job [{job_name}:{batch}] finished but no success - remove from awaiting list, do not add to ready"
-                            )
-                            continue
-                        if job_name in await_finished:
-                            await_finished[job_name].append(batch)
-                        else:
-                            await_finished[job_name] = [batch]
-                jobs_with_params = jobs_with_params_copy
-                expired_sec = int(time.time()) - start_at
-                print(
-                    f"...awaiting continues... seconds left [{TIMEOUT - expired_sec}]"
-                )
-            if await_finished:
-                GHActions.print_in_group(
-                    f"Finished jobs, round [{round_cnt}]:",
-                    [f"{job}:{batches}" for job, batches in await_finished.items()],
-                )
-        GHActions.print_in_group(
-            "Remaining jobs:",
-            [f"{job}:{params['batches']}" for job, params in jobs_with_params.items()],
-        )
-        return await_finished
-
-
-@dataclass
-class CiOptions:
-    # job will be included in the run if any keyword from the list matches job name
-    include_keywords: Optional[List[str]] = None
-    # job will be excluded in the run if any keyword from the list matches job name
-    exclude_keywords: Optional[List[str]] = None
-
-    # list of specified preconfigured ci sets to run
-    ci_sets: Optional[List[str]] = None
-    # list of specified jobs to run
-    ci_jobs: Optional[List[str]] = None
-
-    # batches to run for all multi-batch jobs
-    job_batches: Optional[List[int]] = None
-
-    do_not_test: bool = False
-    no_ci_cache: bool = False
-    upload_all: bool = False
-    no_merge_commit: bool = False
-
-    def as_dict(self) -> Dict[str, Any]:
-        return asdict(self)
-
-    @staticmethod
-    def create_from_run_config(run_config: Dict[str, Any]) -> "CiOptions":
-        return CiOptions(**run_config["ci_options"])
-
-    @staticmethod
-    def create_from_pr_message(
-        debug_message: Optional[str], update_from_api: bool
-    ) -> "CiOptions":
-        """
-        Creates CiOptions instance based on tags found in PR body and/or commit message
-        @commit_message - may be provided directly for debugging purposes, otherwise it will be retrieved from git.
-        """
-        res = CiOptions()
-        pr_info = PRInfo()
-        if (
-            not pr_info.is_pr and not debug_message
-        ):  # if commit_message is provided it's test/debug scenario - do not return
-            # CI options can be configured in PRs only
-            # if debug_message is provided - it's a test
-            return res
-        message = debug_message or GitRunner(set_cwd_to_git_root=True).run(
-            f"{GIT_PREFIX} log {pr_info.sha} --format=%B -n 1"
-        )
-
-        # CI setting example we need to match with re:
-        # - [x] <!---ci_exclude_tsan|msan|ubsan|coverage--> Exclude: All with TSAN, MSAN, UBSAN, Coverage
-        pattern = r"(#|- \[x\] +<!---)([|\w]+)"
-        matches = [match[-1] for match in re.findall(pattern, message)]
-        print(f"CI tags from commit message: [{matches}]")
-
-        if not debug_message:  # to be skipped if debug/test
-            pr_info = PRInfo(
-                pr_event_from_api=update_from_api
-            )  # Fetch updated PR body from GH API
-            matches_pr = [match[-1] for match in re.findall(pattern, pr_info.body)]
-            print(f"CI tags from PR body: [{matches_pr}]")
-            matches = list(set(matches + matches_pr))
-
-            if "do not test" in pr_info.labels:
-                # do_not_test could be set in GH labels
-                res.do_not_test = True
-
-        for match in matches:
-            if match.startswith("job_"):
-                if not res.ci_jobs:
-                    res.ci_jobs = []
-                res.ci_jobs.append(match.removeprefix("job_"))
-            elif match.startswith("ci_set_") and match in CILabels:
-                if not res.ci_sets:
-                    res.ci_sets = []
-                res.ci_sets.append(match)
-            elif match.startswith("ci_include_"):
-                if not res.include_keywords:
-                    res.include_keywords = []
-                res.include_keywords.append(
-                    normalize_check_name(match.removeprefix("ci_include_"))
-                )
-            elif match.startswith("ci_exclude_"):
-                if not res.exclude_keywords:
-                    res.exclude_keywords = []
-                keywords = match.removeprefix("ci_exclude_").split("|")
-                res.exclude_keywords += [
-                    normalize_check_name(keyword) for keyword in keywords
-                ]
-            elif match == CILabels.NO_CI_CACHE:
-                res.no_ci_cache = True
-                print("NOTE: CI Cache will be disabled")
-            elif match == CILabels.UPLOAD_ALL_ARTIFACTS:
-                res.upload_all = True
-                print("NOTE: All binary artifacts will be uploaded")
-            elif match == CILabels.DO_NOT_TEST_LABEL:
-                res.do_not_test = True
-            elif match == CILabels.NO_MERGE_COMMIT:
-                res.no_merge_commit = True
-                print("NOTE: Merge Commit will be disabled")
-            elif match.startswith("batch_"):
-                batches = []
-                try:
-                    batches = [
-                        int(batch) for batch in match.removeprefix("batch_").split("_")
-                    ]
-                except Exception:
-                    print(f"ERROR: failed to parse commit tag [{match}] - skip")
-                if batches:
-                    if not res.job_batches:
-                        res.job_batches = []
-                    res.job_batches += batches
-                    res.job_batches = list(set(res.job_batches))
-            else:
-                print(
-                    f"WARNING: Invalid tag in commit message or PR body [{match}] - skip"
-                )
-
-        return res
-
-    def apply(
-        self,
-        jobs_to_do: List[str],
-        jobs_to_skip: List[str],
-        jobs_params: Dict[str, Dict[str, Any]],
-        pr_info: PRInfo,
-    ) -> Tuple[List[str], List[str], Dict[str, Dict[str, Any]]]:
-        """
-        Applies specified options on CI Run Config
-        Returns updated jobs_to_do, jobs_to_skip, jobs_params
-        """
-        jobs_to_do_requested = []  # type: List[str]
-
-        # -1. Handle "ci_exclude_" tags if any
-        if self.exclude_keywords:
-            new_jobs_to_do = list(jobs_to_do)
-            for job in jobs_to_do:
-                found = False
-                for keyword in self.exclude_keywords:
-                    if keyword in normalize_check_name(job):
-                        print(
-                            f"Job [{job}] matches Exclude keyword [{keyword}] - remove"
-                        )
-                        found = True
-                        break
-                if found:
-                    new_jobs_to_do.remove(job)
-            jobs_to_do = new_jobs_to_do
-
-        # 0. Handle "ci_include_" tags if any
-        if self.include_keywords:
-            for job in jobs_to_do:
-                found = False
-                for keyword in self.include_keywords:
-                    if keyword in normalize_check_name(job):
-                        print(f"Job [{job}] matches Include keyword [{keyword}] - add")
-                        found = True
-                        break
-                if found:
-                    job_with_parents = CI_CONFIG.get_job_with_parents(job)
-                    for job in job_with_parents:
-                        if job in jobs_to_do and job not in jobs_to_do_requested:
-                            jobs_to_do_requested.append(job)
-            if not jobs_to_do_requested:
-                print(
-                    f"WARNING: Include tags are set but no job configured - Invalid tags, probably [{self.include_keywords}]"
-                )
-            if JobNames.STYLE_CHECK not in jobs_to_do_requested:
-                # Style check must not be omitted
-                jobs_to_do_requested.append(JobNames.STYLE_CHECK)
-
-        # 1. Handle "ci_set_" tags if any
-        if self.ci_sets:
-            for tag in self.ci_sets:
-                label_config = CI_CONFIG.get_label_config(tag)
-                assert label_config, f"Unknown tag [{tag}]"
-                print(
-                    f"NOTE: CI Set's tag: [{tag}], add jobs: [{label_config.run_jobs}]"
-                )
-                # match against @jobs_to_do and @jobs_to_skip to remove non-relevant entries from @label_config.run_jobs
-                jobs_to_do_requested += [
-                    job
-                    for job in label_config.run_jobs
-                    if job in jobs_to_do or job in jobs_to_skip
-                ]
-
-        # FIXME: to be removed in favor of include/exclude
-        # 2. Handle "job_" tags if any
-        if self.ci_jobs:
-            for job in self.ci_jobs:
-                job_with_parents = CI_CONFIG.get_job_with_parents(job)
-                print(
-                    f"NOTE: CI Job's tag: [#job_{job}], add jobs: [{job_with_parents}]"
-                )
-                # always add requested job itself, even if it could be skipped
-                jobs_to_do_requested.append(job_with_parents[0])
-                for parent in job_with_parents[1:]:
-                    if parent in jobs_to_do and parent not in jobs_to_do_requested:
-                        jobs_to_do_requested.append(parent)
-
-        # 3. Handle "do not test"
-        if self.do_not_test:
-            label_config = CI_CONFIG.get_label_config(CILabels.DO_NOT_TEST_LABEL)
-            assert label_config
-            print(
-                f"NOTE: CI 'do not test' setting applied, set jobs: [{label_config.run_jobs}]"
-            )
-            if jobs_to_do_requested:
-                print(
-                    "WARNING: 'do not test' is used alongside with other CI modifying tags - 'do not test' prevails"
-                )
-            jobs_to_do_requested = list(label_config.run_jobs)
-
-        if jobs_to_do_requested:
-            jobs_to_do_requested = list(set(jobs_to_do_requested))
-            print(
-                f"NOTE: Only specific job(s) were requested by user's input: [{jobs_to_do_requested}]"
-            )
-            jobs_to_do = list(
-                set(job for job in jobs_to_do_requested if job not in jobs_to_skip)
-            )
-            # if requested job does not have params in jobs_params (it happens for "run_by_label" job)
-            #   we need to add params - otherwise it won't run as "batches" list will be empty
-            for job in jobs_to_do:
-                if job not in jobs_params:
-                    job_config = CI_CONFIG.get_job_config(job)
-                    num_batches = job_config.num_batches
-                    jobs_params[job] = {
-                        "batches": list(range(num_batches)),
-                        "num_batches": num_batches,
-                        "run_by_ci_option": job_config.run_by_ci_option
-                        and pr_info.is_pr,
-                    }
-
-        # 4. Handle "batch_" tags
-        if self.job_batches:
-            print(
-                f"NOTE: Only specific job batches were requested [{self.job_batches}]"
-            )
-            for job, params in jobs_params.items():
-                if params["num_batches"] > 1:
-                    params["batches"] = self.job_batches
-
-        for job in jobs_to_do[:]:
-            job_param = jobs_params[job]
-            if job_param["run_by_ci_option"] and job not in jobs_to_do_requested:
-                print(
-                    f"Erasing job '{job}' from list because it's not in included set, but will run only by include"
-                )
-                jobs_to_skip.append(job)
-                jobs_to_do.remove(job)
-
-        return jobs_to_do, jobs_to_skip, jobs_params
 
 
 def get_check_name(check_name: str, batch: int, num_batches: int) -> str:
     res = check_name
     if num_batches > 1:
         res = f"{check_name} [{batch+1}/{num_batches}]"
-    return res
-
-
-def normalize_check_name(check_name: str) -> str:
-    res = check_name.lower()
-    for r in ((" ", "_"), ("(", "_"), (")", "_"), (",", "_"), ("/", "_")):
-        res = res.replace(*r)
     return res
 
 
@@ -1279,12 +337,12 @@ def _mark_success_action(
             # ci_cache.delete_pending(job, batch, num_batches, release_branch=True)
         if job_status and job_status.is_ok():
             ci_cache.push_successful(
-                job, batch, num_batches, job_status, pr_info.is_release_branch
+                job, batch, num_batches, job_status, pr_info.is_release
             )
             print(f"Job [{job}] is ok")
         elif job_status and not job_status.is_ok():
             ci_cache.push_failed(
-                job, batch, num_batches, job_status, pr_info.is_release_branch
+                job, batch, num_batches, job_status, pr_info.is_release
             )
             print(f"Job [{job}] is failed with status [{job_status.status}]")
         else:
@@ -1292,7 +350,7 @@ def _mark_success_action(
                 description="dummy description", status=ERROR, report_url="dummy url"
             )
             ci_cache.push_failed(
-                job, batch, num_batches, job_status, pr_info.is_release_branch
+                job, batch, num_batches, job_status, pr_info.is_release
             )
             print(f"No CommitStatusData for [{job}], push dummy failure to ci_cache")
 
@@ -1362,170 +420,38 @@ def _configure_docker_jobs(docker_digest_or_latest: bool) -> Dict:
 
 
 def _configure_jobs(
-    job_digester: JobDigester,
     s3: S3Helper,
     pr_info: PRInfo,
-    ci_options: CiOptions,
-) -> Dict:
-    ## a. digest each item from the config
-    job_digester = JobDigester()
-    jobs_params: Dict[str, Dict] = {}
-    jobs_to_do: List[str] = []
-    jobs_to_skip: List[str] = []
-    digests: Dict[str, str] = {}
+    ci_settings: CiSettings,
+) -> CiCache:
+    """
+    returns CICache instance with configured job's data
+    :param s3:
+    :param pr_info:
+    :param ci_settings:
+    :return:
+    """
 
-    # FIXME: find better place for these config variables
-    DOCS_CHECK_JOBS = [JobNames.DOCS_CHECK, JobNames.STYLE_CHECK]
-    MQ_JOBS = [
-        JobNames.STYLE_CHECK,
-        JobNames.FAST_TEST,
-        Build.BINARY_RELEASE,
-        JobNames.UNIT_TEST,
-    ]
-    # Must always calculate digest for these jobs for CI Cache to function (they define s3 paths where records are stored)
-    REQUIRED_DIGESTS = [JobNames.DOCS_CHECK, Build.PACKAGE_RELEASE]
-    if pr_info.has_changes_in_documentation_only():
-        print(f"WARNING: Only docs are changed - will run only [{DOCS_CHECK_JOBS}]")
-    if pr_info.is_merge_queue:
-        print(f"WARNING: It's a MQ run - will run only [{MQ_JOBS}]")
-
-    print("::group::Job Digests")
-    for job in CI_CONFIG.job_generator(pr_info.head_ref if CI else "dummy_branch_name"):
-        if (
-            pr_info.is_merge_queue
-            and job not in MQ_JOBS
-            and job not in REQUIRED_DIGESTS
-        ):
-            # We still need digest for JobNames.DOCS_CHECK since CiCache depends on it (FIXME)
-            continue
-        if pr_info.is_master and job in MQ_JOBS:
-            # On master - skip jobs that run in MQ
-            continue
-        if (
-            pr_info.has_changes_in_documentation_only()
-            and job not in DOCS_CHECK_JOBS
-            and job not in REQUIRED_DIGESTS
-        ):
-            continue
-        digest = job_digester.get_job_digest(CI_CONFIG.get_digest_config(job))
-        digests[job] = digest
-        print(f"    job [{job.rjust(50)}] has digest [{digest}]")
-    print("::endgroup::")
-
-    ## b. check what we need to run
-    ci_cache = None
-    if not ci_options.no_ci_cache and CI:
-        ci_cache = CiCache(s3, digests).update()
-        ci_cache.print_status()
-
-    jobs_to_wait: Dict[str, Dict[str, Any]] = {}
-    randomization_buckets = {}  # type: Dict[str, Set[str]]
-
-    for job, digest in digests.items():
-        job_config = CI_CONFIG.get_job_config(job)
-        num_batches: int = job_config.num_batches
-        batches_to_do: List[int] = []
-        add_to_skip = False
-
-        if pr_info.is_merge_queue and job not in MQ_JOBS:
-            continue
-        if job_config.pr_only and pr_info.is_release_branch:
-            continue
-        if (
-            job_config.release_only
-            and not job_config.run_by_ci_option
-            and not pr_info.is_release_branch
-        ):
-            continue
-
-        # fill job randomization buckets (for jobs with configured @random_bucket property)
-        if job_config.random_bucket:
-            if not job_config.random_bucket in randomization_buckets:
-                randomization_buckets[job_config.random_bucket] = set()
-            randomization_buckets[job_config.random_bucket].add(job)
-
-        for batch in range(num_batches):  # type: ignore
-            if job_config.run_by_label:
-                # this job controlled by label, add to todo if its label is set in pr
-                if job_config.run_by_label in pr_info.labels:
-                    batches_to_do.append(batch)
-            elif job_config.run_always:
-                # always add to todo
-                batches_to_do.append(batch)
-            elif not ci_cache:
-                batches_to_do.append(batch)
-            elif not ci_cache.is_successful(
-                job,
-                batch,
-                num_batches,
-                release_branch=pr_info.is_release_branch
-                and job_config.required_on_release_branch,
-            ):
-                # ci cache is enabled and job is not in the cache - add
-                batches_to_do.append(batch)
-
-                # check if it's pending in the cache
-                if ci_cache.is_pending(
-                    job,
-                    batch,
-                    num_batches,
-                    release_branch=pr_info.is_release_branch
-                    and job_config.required_on_release_branch,
-                ):
-                    if job in jobs_to_wait:
-                        jobs_to_wait[job]["batches"].append(batch)
-                    else:
-                        jobs_to_wait[job] = {
-                            "batches": [batch],
-                            "num_batches": num_batches,
-                        }
-            else:
-                add_to_skip = True
-
-        if batches_to_do:
-            jobs_to_do.append(job)
-            jobs_params[job] = {
-                "batches": batches_to_do,
-                "num_batches": num_batches,
-                "run_by_ci_option": job_config.run_by_ci_option and pr_info.is_pr,
-            }
-        elif add_to_skip:
-            # treat job as being skipped only if it's controlled by digest
-            jobs_to_skip.append(job)
-
-    if not pr_info.is_release_branch:
-        # randomization bucket filtering (pick one random job from each bucket, for jobs with configured random_bucket property)
-        for _, jobs in randomization_buckets.items():
-            jobs_to_remove_randomization = set()
-            bucket_ = list(jobs)
-            random.shuffle(bucket_)
-            while len(bucket_) > 1:
-                random_job = bucket_.pop()
-                if random_job in jobs_to_do:
-                    jobs_to_remove_randomization.add(random_job)
-            if jobs_to_remove_randomization:
-                print(
-                    f"Following jobs will be removed due to randomization bucket: [{jobs_to_remove_randomization}]"
-                )
-                jobs_to_do = [
-                    job for job in jobs_to_do if job not in jobs_to_remove_randomization
-                ]
-
-    jobs_to_do, jobs_to_skip, jobs_params = ci_options.apply(
-        jobs_to_do, jobs_to_skip, jobs_params, pr_info
+    # get all jobs
+    job_configs = CI_CONFIG.get_workflow_jobs_with_configs(
+        is_mq=pr_info.is_merge_queue,
+        is_docs_only=pr_info.has_changes_in_documentation_only(),
+        is_master=pr_info.is_master,
     )
 
-    return {
-        "digests": digests,
-        "jobs_to_do": jobs_to_do,
-        "jobs_to_skip": jobs_to_skip,
-        "jobs_to_wait": {
-            job: params for job, params in jobs_to_wait.items() if job in jobs_to_do
-        },
-        "jobs_params": {
-            job: params for job, params in jobs_params.items() if job in jobs_to_do
-        },
-    }
+    # filter jobs in accordance with ci settings
+    job_configs = ci_settings.apply(
+        job_configs, pr_info.is_release, is_pr=pr_info.is_pr, labels=pr_info.labels
+    )
+
+    # check jobs in ci cache
+    ci_cache = CiCache.calc_digests_and_create(
+        s3, job_configs, cache_enabled=not ci_settings.no_ci_cache and CI
+    )
+    ci_cache.update()
+    ci_cache.apply(job_configs, is_release=pr_info.is_release)
+
+    return ci_cache
 
 
 def _generate_ci_stage_config(jobs_data: Dict[str, Any]) -> Dict[str, Dict[str, Any]]:
@@ -1537,7 +463,7 @@ def _generate_ci_stage_config(jobs_data: Dict[str, Any]) -> Dict[str, Dict[str, 
     """
     result = {}  # type: Dict[str, Any]
     stages_to_do = []
-    for job in jobs_data["jobs_to_do"]:
+    for job in jobs_data:
         stage_type = CI_CONFIG.get_job_ci_stage(job)
         if stage_type == CIStages.NA:
             continue
@@ -1568,7 +494,7 @@ def _create_gh_status(
 
 
 def _update_gh_statuses_action(indata: Dict, s3: S3Helper) -> None:
-    if CiOptions.create_from_run_config(indata).no_ci_cache:
+    if CiSettings.create_from_run_config(indata).no_ci_cache:
         print("CI cache is disabled - skip restoring commit statuses from CI cache")
         return
     job_digests = indata["jobs_data"]["digests"]
@@ -2024,16 +950,13 @@ def main() -> int:
             meta.run_id = int(GITHUB_RUN_ID)
             meta.push_meta()
 
-        ci_options = CiOptions.create_from_pr_message(
+        ci_settings = CiSettings.create_from_pr_message(
             args.commit_message or None, update_from_api=True
         )
 
-        # tokens = _fetch_commit_tokens(message, pr_info)
-        if ci_options.no_merge_commit and CI:
+        if ci_settings.no_merge_commit and CI:
             git_runner.run(f"{GIT_PREFIX} checkout {pr_info.sha}")
-            git_ref = git_runner.run(f"{GIT_PREFIX} rev-parse HEAD")
 
-        docker_data = {}
         git_ref = git_runner.run(f"{GIT_PREFIX} rev-parse HEAD")
 
         # let's get CH version
@@ -2046,72 +969,37 @@ def main() -> int:
             else {}
         )
 
-        job_digester = JobDigester()
-        build_digest = job_digester.get_job_digest(
-            CI_CONFIG.get_digest_config("package_release")
+        ci_cache = _configure_jobs(
+            s3,
+            pr_info,
+            ci_settings,
         )
-        docs_digest = job_digester.get_job_digest(
-            CI_CONFIG.get_digest_config(JobNames.DOCS_CHECK)
-        )
-        jobs_data = (
-            _configure_jobs(
-                job_digester,
-                s3,
-                pr_info,
-                ci_options,
-            )
-            if not args.skip_jobs
-            else {}
-        )
+        ci_cache.print_status()
 
-        if not args.skip_jobs:
-            ci_cache = CiCache(s3, jobs_data["digests"])
+        if CI and not pr_info.is_merge_queue:
+            # wait for pending jobs to be finished, await_jobs is a long blocking call
+            ci_cache.await_pending_jobs(pr_info.is_release)
 
-            if pr_info.is_master:
-                # wait for pending jobs to be finished, await_jobs is a long blocking call
-                # wait pending jobs (for now only on release/master branches)
-                ready_jobs_batches_dict = ci_cache.await_jobs(
-                    jobs_data.get("jobs_to_wait", {}), pr_info.is_release_branch
-                )
-                jobs_to_do = jobs_data["jobs_to_do"]
-                jobs_to_skip = jobs_data["jobs_to_skip"]
-                jobs_params = jobs_data["jobs_params"]
-                for job, batches in ready_jobs_batches_dict.items():
-                    if job not in jobs_params:
-                        print(f"WARNING: Job [{job}] is not in the params list")
-                        continue
-                    for batch in batches:
-                        jobs_params[job]["batches"].remove(batch)
-                    if not jobs_params[job]["batches"]:
-                        jobs_to_do.remove(job)
-                        jobs_to_skip.append(job)
-                        del jobs_params[job]
-
-                # set planned jobs as in-progress in CI cache
-                for job in jobs_data["jobs_to_do"]:
-                    config = CI_CONFIG.get_job_config(job)
-                    if config.run_always or config.run_by_label:
-                        continue
-                    job_params = jobs_data["jobs_params"][job]
-                    ci_cache.push_pending(
-                        job,
-                        job_params["batches"],
-                        config.num_batches,
-                        release_branch=pr_info.is_release_branch,
-                    )
-
-            if "jobs_to_wait" in jobs_data:
-                del jobs_data["jobs_to_wait"]
+            if pr_info.is_release:
+                ci_cache.push_pending_all(pr_info.is_release)
 
         # conclude results
         result["git_ref"] = git_ref
         result["version"] = version
-        result["build"] = build_digest
-        result["docs"] = docs_digest
-        result["ci_options"] = ci_options.as_dict()
+        result["build"] = ci_cache.job_digests[Build.PACKAGE_RELEASE]
+        result["docs"] = ci_cache.job_digests[JobNames.DOCS_CHECK]
+        result["ci_settings"] = ci_settings.as_dict()
         if not args.skip_jobs:
-            result["stages_data"] = _generate_ci_stage_config(jobs_data)
-        result["jobs_data"] = jobs_data
+            result["stages_data"] = _generate_ci_stage_config(ci_cache.jobs_to_do)
+        result["jobs_data"] = {
+            "jobs_to_do": list(ci_cache.jobs_to_do),
+            "jobs_to_skip": ci_cache.jobs_to_skip,
+            "digests": ci_cache.job_digests,
+            "jobs_params": {
+                job: {"batches": config.batches, "num_batches": config.num_batches}
+                for job, config in ci_cache.jobs_to_do.items()
+            },
+        }
         result["docker_data"] = docker_data
     ### CONFIGURE action: end
 
@@ -2123,7 +1011,7 @@ def main() -> int:
     ### RUN action: start
     elif args.run:
         assert indata
-        ci_options = CiOptions.create_from_run_config(indata)
+        ci_settings = CiSettings.create_from_run_config(indata)
         check_name = args.job_name
         check_name_with_group = _get_ext_check_name(check_name)
         print(
@@ -2131,10 +1019,10 @@ def main() -> int:
         )
         previous_status = None
         if CI_CONFIG.is_build_job(check_name):
-            # this is a build job - check if build report is present
+            # this is a build job - check if a build report is present
             build_result = (
                 BuildResult.load_any(check_name, pr_info.number, pr_info.head_ref)
-                if not ci_options.no_ci_cache
+                if not ci_settings.no_ci_cache
                 else None
             )
             if build_result:
@@ -2172,7 +1060,7 @@ def main() -> int:
                     print("::endgroup::")
 
             # ci cache check
-            if not previous_status and not ci_options.no_ci_cache:
+            if not previous_status and not ci_settings.no_ci_cache:
                 ci_cache = CiCache(s3, indata["jobs_data"]["digests"]).update()
                 job_config = CI_CONFIG.get_job_config(check_name)
                 if ci_cache.is_successful(
@@ -2224,7 +1112,7 @@ def main() -> int:
                     not pr_info.is_pr
                     or args.job_name
                     not in CI_CONFIG.get_builds_for_report(JobNames.BUILD_CHECK_SPECIAL)
-                    or CiOptions.create_from_run_config(indata).upload_all
+                    or CiSettings.create_from_run_config(indata).upload_all
                 )
 
                 build_name = args.job_name
