+++ tests/ci/test_ci_options.py
@@ -3,8 +3,8 @@
 # type: ignore
 
 import unittest
-from ci import CiOptions
-from pr_info import PRInfo
+from ci_settings import CiSettings
+from ci_config import JobConfig
 
 _TEST_BODY_1 = """
 #### Run only:
@@ -54,6 +54,14 @@ _TEST_JOB_LIST = [
     "Fast test",
     "package_release",
     "package_asan",
+    "package_aarch64",
+    "package_release_coverage",
+    "package_debug",
+    "package_tsan",
+    "package_msan",
+    "package_ubsan",
+    "binary_release",
+    "fuzzers",
     "Docker server image",
     "Docker keeper image",
     "Install packages (amd64)",
@@ -129,10 +137,12 @@ _TEST_JOB_LIST = [
     "Bugfix validation",
 ]
 
+_TEST_JOB_LIST_2 = ["Style check", "Fast test", "fuzzers"]
+
 
 class TestCIOptions(unittest.TestCase):
     def test_pr_body_parsing(self):
-        ci_options = CiOptions.create_from_pr_message(
+        ci_options = CiSettings.create_from_pr_message(
             _TEST_BODY_1, update_from_api=False
         )
         self.assertFalse(ci_options.do_not_test)
@@ -144,7 +154,7 @@ class TestCIOptions(unittest.TestCase):
 
     def test_options_applied(self):
         self.maxDiff = None
-        ci_options = CiOptions.create_from_pr_message(
+        ci_options = CiSettings.create_from_pr_message(
             _TEST_BODY_2, update_from_api=False
         )
         self.assertCountEqual(
@@ -155,24 +165,33 @@ class TestCIOptions(unittest.TestCase):
             ci_options.exclude_keywords,
             ["tsan", "foobar", "aarch64", "analyzer", "s3_storage", "coverage"],
         )
-        jobs_to_do = list(_TEST_JOB_LIST)
-        jobs_to_skip = []
-        job_params = {
-            "Stateless tests (azure, asan)": {
-                "batches": list(range(3)),
-                "num_batches": 3,
-                "run_by_ci_option": True,
-            }
-        }
-        jobs_to_do, jobs_to_skip, job_params = ci_options.apply(
-            jobs_to_do, jobs_to_skip, job_params, PRInfo()
+
+        jobs_configs = {job: JobConfig() for job in _TEST_JOB_LIST}
+        jobs_configs[
+            "fuzzers"
+        ].run_by_label = (
+            "TEST_LABEL"  # check "fuzzers" appears in the result due to the label
+        )
+        jobs_configs[
+            "Integration tests (asan)"
+        ].release_only = (
+            True  # still must be included as it's set with include keywords
+        )
+        filtered_jobs = list(
+            ci_options.apply(
+                jobs_configs, is_release=False, is_pr=True, labels=["TEST_LABEL"]
+            )
         )
         self.assertCountEqual(
-            jobs_to_do,
+            filtered_jobs,
             [
                 "Style check",
+                "fuzzers",
                 "package_release",
                 "package_asan",
+                "package_debug",
+                "package_msan",
+                "package_ubsan",
                 "Stateless tests (asan)",
                 "Stateless tests (azure, asan)",
                 "Stateless tests flaky check (asan)",
@@ -187,54 +206,88 @@ class TestCIOptions(unittest.TestCase):
         )
 
     def test_options_applied_2(self):
+        jobs_configs = {job: JobConfig() for job in _TEST_JOB_LIST_2}
+        jobs_configs["Style check"].release_only = True
+        jobs_configs["Fast test"].pr_only = True
+        jobs_configs["fuzzers"].run_by_label = "TEST_LABEL"
+        # no settings are set
+        filtered_jobs = list(
+            CiSettings().apply(jobs_configs, is_release=False, is_pr=True, labels=[])
+        )
+        self.assertCountEqual(
+            filtered_jobs,
+            [
+                "Fast test",
+            ],
+        )
+
+        filtered_jobs = list(
+            CiSettings().apply(jobs_configs, is_release=True, is_pr=False, labels=[])
+        )
+        self.assertCountEqual(
+            filtered_jobs,
+            [
+                "Style check",
+            ],
+        )
+
+    def test_options_applied_3(self):
+        ci_settings = CiSettings()
+        ci_settings.include_keywords = ["Style"]
+        jobs_configs = {job: JobConfig() for job in _TEST_JOB_LIST_2}
+        jobs_configs["Style check"].release_only = True
+        jobs_configs["Fast test"].pr_only = True
+        # no settings are set
+        filtered_jobs = list(
+            ci_settings.apply(
+                jobs_configs, is_release=False, is_pr=True, labels=["TEST_LABEL"]
+            )
+        )
+        self.assertCountEqual(
+            filtered_jobs,
+            [
+                "Style check",
+            ],
+        )
+
+        ci_settings.include_keywords = ["Fast"]
+        filtered_jobs = list(
+            ci_settings.apply(
+                jobs_configs, is_release=True, is_pr=False, labels=["TEST_LABEL"]
+            )
+        )
+        self.assertCountEqual(
+            filtered_jobs,
+            [
+                "Style check",
+            ],
+        )
+
+    def test_options_applied_4(self):
         self.maxDiff = None
-        ci_options = CiOptions.create_from_pr_message(
+        ci_options = CiSettings.create_from_pr_message(
             _TEST_BODY_3, update_from_api=False
         )
         self.assertCountEqual(ci_options.include_keywords, ["analyzer"])
         self.assertIsNone(ci_options.exclude_keywords)
-        jobs_to_do = list(_TEST_JOB_LIST)
-        jobs_to_skip = []
-        job_params = {}
-        jobs_to_do, jobs_to_skip, job_params = ci_options.apply(
-            jobs_to_do, jobs_to_skip, job_params, PRInfo()
+        jobs_configs = {job: JobConfig() for job in _TEST_JOB_LIST}
+        jobs_configs[
+            "fuzzers"
+        ].run_by_label = "TEST_LABEL"  # check "fuzzers" does not appears in the result
+        jobs_configs["Integration tests (asan)"].release_only = True
+        filtered_jobs = list(
+            ci_options.apply(
+                jobs_configs, is_release=False, is_pr=True, labels=["TEST_LABEL"]
+            )
         )
         self.assertCountEqual(
-            jobs_to_do,
+            filtered_jobs,
             [
                 "Style check",
                 "Integration tests (asan, old analyzer)",
                 "package_release",
                 "Stateless tests (release, old analyzer, s3, DatabaseReplicated)",
                 "package_asan",
+                "fuzzers",
             ],
-        )
-
-    def test_options_applied_3(self):
-        self.maxDiff = None
-        ci_options = CiOptions.create_from_pr_message(
-            _TEST_BODY_4, update_from_api=False
-        )
-        self.assertIsNone(ci_options.include_keywords, None)
-        self.assertIsNone(ci_options.exclude_keywords, None)
-        jobs_to_do = list(_TEST_JOB_LIST)
-        jobs_to_skip = []
-        job_params = {}
-
-        for job in _TEST_JOB_LIST:
-            if "Stateless" in job:
-                job_params[job] = {
-                    "batches": list(range(3)),
-                    "num_batches": 3,
-                    "run_by_ci_option": "azure" in job,
-                }
-            else:
-                job_params[job] = {"run_by_ci_option": False}
-
-        jobs_to_do, jobs_to_skip, job_params = ci_options.apply(
-            jobs_to_do, jobs_to_skip, job_params, PRInfo()
-        )
-        self.assertNotIn(
-            "Stateless tests (azure, asan)",
-            jobs_to_do,
         )
